requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
packaging==24.0
panda==0.3.1
pipdeptree==2.19.1
protobuf==5.26.1
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
python-decouple==3.8
requests==2.31.0
rest-framework-simplejwt==0.0.2
setuptools==68.2.2
sqlparse==0.5.0
timedelta==2020.12.3
urllib3==2.2.1
wheel==0.41.2
